Fix partially obstructed tiles by bottom-positioned page menu (#100)
When there is enough content to enable vertical scrolling and page menu toolbar
is at the bottom, the tiles that were at the bottom might have been obstructed
by the toolbar, making it impossible to click some parts of them.

Fixed by adding bottom padding to page area equal to size of the toolbar
(+ toolbar's bottom offset when not 0). This ensures that tiles can be
scrolled fully in view without being obstructed.