Use padding in bottom-positioned pages-menu instead of bottom offset
The only reason for this change is to prepare code for future fix.
It should be almost equally visually to previous code.

Previous code applied bottom offset to bottom-positioned page menu which
created a gap that would let clicks go through it if page was scrolled.
Using padding instead makes sure there is no gap there and also allows
to created better scrolling indicator in the future.