Use `field` mechanism for value in sliders
Just like the LightSliderConf, allow the configuration to include a `field` and use that as the attribute name to be used for the value of the slider.

This is the same the is being done in `getLightSliderConf`, so seems suitable in that context, if I understood everything correctly. This was my approach to solve #209, but I may have overlooked "the proper way", maybe?