feat: support swiping to switch pages
Use hammer library to add support for swiping to switch to next/previous
page.

When menu is on the left, changing pages is done using swipe up/down
motion and when on the bottom, using swipe left/right motion.

With menu on the bottom, for page toggle animation to match swipe
direction, it was necessary to change how pages are layouted (needed to
be horizontally instead of vertically).

angular-long-press library was replaced with hammer as it supports the
same and was needed for swiping anyway.

Fixes #242