Fix item's state behavior
This alters the `item.state(s)` fallback logic, in order for the following to work:

```
state: function (item, entity) {
   return entity.attributes.can_cancel ? entity.state : null;
},
```

The purpose is to remove the state field from tiles of one-shot scripts, which have no meaningful state.

The new fallback falls in line with the 'highly customizable' objective of TileBoard:
* if the user provides `state` or `states`, the field is used - no matter what, so also `null`, `false`, `{}` or similar results.
* if the user provides sth, which is not a function, string or object, it is simply passed on as the state.
* if none of the fields are provided, the `entity.state` is returned.

I believe this is very similar to the old logic, but more consistent and more customizable. However, people might have relied on the old behavior; e.g. (whyever) providing `state` and `states` together using the latter as a fallback for empty result of the former.

For you to decide, if this change is useful. I believe it is - for the sake of consistency and customizablility. Hence the PR. ;-)