Fix chart overflowing when width to height ratio becomes over 1.3 or so
Shortly after screen width starts becoming bigger than the height, the
chart would overflow the screen vertically. That's because container
was sized using padding-bottom: 75% which means 75% of container width
and that would become bigger than available space at some point. There
is no easy CSS solution to handle "75% of width but max 100% height", but
if we just use 100% when in landscape mode (screen ratio with width
bigger than height) then it's impossible to trigger the problem and it
gives better viewing area for the chart also.

Resolves #296