Hide browser UI on iOS 13 ("Add to Home Screen") (#364)
iOS 13 requires not only the "apple-mobile-web-app-capable" meta tag but also either one of the entries "display": "standalone" or "display": "fullscreen" in the manifest file. Otherwise the "Add to Home Screen" function only creates a link to TileBoard which opens the interface in Safari. With the manifest "display" entry it behaves more like a "progressive web app", hiding the browser UI and address bar.