Fix for sun.sun as requested
You need to pass as a property

sun: {
  id: 'your sun entity',
  night: 'night state of sun entity'
}

Checking if sun.sun exists with $scope.getItemEntity will open a noty which is not desirable here. I could be missing something as well. My understanding is that hassio entities are not global within Tileboard (again, I might be missing something). 

This might be for another thread but I have actually forced the entities into CONFIG as I needed to easy access them. I have not mentioned that nor submitted because I am not sure if this is a good practice. You guys let me know.

Cheers