transparent background in iframe tile
this allows to set the background color in the popup, which is otherweise impossible.
this might change a few people's background in regular iframe tiles, if they have kept the default background color.
(which I doubt, first because it's quite grey, second because the page in the iframe probably has its own background color...)