fix: cache data in item instead of entity.attributes
this change also makes this the default for a few other funtions, namely sliders.

fixes #425

BREAKING CHANGE: The order of attrs and def is changed for a few sliders. This should not normally be a problem, because user would only use one of the possibilities. But... just to give note on that.