fix(LIGHT): show color picker based on feature check
This is a behavior change since now even if the user doesn't define
sliders and light supports color change then picker will still show up
on long press.

We maybe should leave support for "item.colorpicker" to be able to
override it to "false" but not sure if it's worth it.