fix(CAMERA_STREAM): stream fails to resume after being suspended
Appears to be a regression in v1 of hls.js. An upstream issue created
at https://github.com/video-dev/hls.js/issues/3732

But I also found similar issue (https://github.com/video-dev/hls.js/issues/2473)
that seems to suggest that it's best to just re-initialize the Hls
instance instead of re-attaching the media so I'll just go with that
for now.

Fixes #688