fix(HISTORY): fix localized tooltip time
Apparently, the new localization features convert the `label` of the `tooltipItem` to something local.
This results in something like `Mai 23, 2021, 1:03:51`, for me in German.
If we try to convert that using `timeAgo`, it fails.
The fix uses the raw `dataset[].data[].x`, which is `Sun May 23 2021 01:03:51 GMT+0200 (Mitteleuropäische Sommerzeit)` in my case.
This value can be converted to a time ago properly.